java-micro-master/finalhunger 2/ecom/src/main/java/com/hungerhub/services/admin/category/CategoryServiceImpl.java
package com.hungerhub.services.admin.category;

import com.hungerhub.dto.CategoryDto;
import com.hungerhub.entity.Category;
import com.hungerhub.repository.CategoryRepository;

import lombok.RequiredArgsConstructor;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class CategoryServiceImpl implements CategoryService{
	@Autowired
    private  CategoryRepository categoryRepository;

    public Category createcategory(CategoryDto categoryDto){
        Category category = new Category();
        category.setName(categoryDto.getName());
        category.setDescription(categoryDto.getDescription());

        return categoryRepository.save(category);
    }

    public List<Category> getAllCategories(){
        return categoryRepository.findAll();
    }
}
